pyproject.toml
[project]
name = "target-lock"
version = "0.1.0"
description = "Back-projection-based open-loop target locking controllers with lockon simulator integration."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "numpy<2.3",
]

[project.optional-dependencies]
sim = [
    "grpcio>=1.78.0,<2.0.0",
    "protobuf>=7.34.0,<8.0.0",
    "opencv-python>=4.13.0.92,<5.0.0.0",
]
dev = [
    "pytest>=8.3.0,<9.0.0",
]

[project.scripts]
target-lock-static = "target_lock.cli.static_open_loop:main"
target-lock-move = "target_lock.cli.move_open_loop:main"
target-lock-square-pid = "target_lock.cli.square_pid:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
